# api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoCertX Control Plane and Agent API",
    "version": "1.0.1-t03",
    "description": "Initial frozen API contract for GA phase one. Covers control plane REST APIs, agent REST protocol, and shared error semantics."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/audit-events": {
      "get": {
        "tags": ["audit"],
        "operationId": "listAuditEvents",
        "summary": "List audit events",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Audit event list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/audit-events/{id}": {
      "get": {
        "tags": ["audit"],
        "operationId": "getAuditEvent",
        "summary": "Get audit event detail",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": {
          "200": {
            "description": "Audit event detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/audit-events/export-csv": {
      "post": {
        "tags": ["audit"],
        "operationId": "exportAuditEventsCSV",
        "summary": "Export audit events as CSV",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuditExportRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audit CSV export",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/api/v1/settings/webhooks": {
      "get": {
        "tags": ["settings"],
        "operationId": "listWebhookEndpoints",
        "summary": "List webhook endpoints",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Webhook endpoint list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "post": {
        "tags": ["settings"],
        "operationId": "createWebhookEndpoint",
        "summary": "Create webhook endpoint",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebhookUpsertRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created webhook endpoint",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/settings/webhooks/{id}": {
      "put": {
        "tags": ["settings"],
        "operationId": "updateWebhookEndpoint",
        "summary": "Update webhook endpoint",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebhookUpsertRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated webhook endpoint",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/settings/renewal-window": {
      "get": {
        "tags": ["settings"],
        "operationId": "getRenewalWindowSettings",
        "summary": "Get renewal window settings",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Renewal window settings",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "put": {
        "tags": ["settings"],
        "operationId": "updateRenewalWindowSettings",
        "summary": "Update renewal window settings",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RenewalWindowSettingsRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated renewal window settings",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
    "/api/v1/settings/security": {
      "get": {
        "tags": ["settings"],
        "operationId": "getSecuritySettings",
        "summary": "Get security settings",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Security settings",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "put": {
        "tags": ["settings"],
        "operationId": "updateSecuritySettings",
        "summary": "Update security settings",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SecuritySettingsRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated security settings",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ObjectEnvelope" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["request_id", "error"],
        "properties": {
          "request_id": { "type": "string" },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ErrorDetail" }
              }
            }
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "field": { "type": "string" },
          "reason": { "type": "string" }
        }
      },
      "PageMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "minimum": 1 },
          "page_size": { "type": "integer", "minimum": 1 },
          "total": { "type": "integer", "minimum": 0 }
        }
      },
      "ObjectEnvelope": {
        "type": "object",
        "required": ["request_id", "data"],
        "properties": {
          "request_id": { "type": "string" },
          "data": { "type": "object" }
        }
      },
      "ListEnvelope": {
        "type": "object",
        "required": ["request_id", "items"],
        "properties": {
          "request_id": { "type": "string" },
          "items": {
            "type": "array",
            "items": { "type": "object" }
          },
          "meta": { "$ref": "#/components/schemas/PageMeta" }
        }
      },
...
      "WebhookUpsertRequest": {
        "type": "object",
        "required": ["name", "url", "event_types"],
        "properties": {
          "name": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "secret": { "type": "string" },
          "event_types": {
            "type": "array",
            "items": { "type": "string" }
          },
          "enabled": { "type": "boolean" }
        }
      },
      "RenewalWindowSettingsRequest": {
        "type": "object",
        "required": ["days_before_expiry"],
        "properties": {
          "days_before_expiry": { "type": "integer", "minimum": 1 },
          "scan_interval_minutes": { "type": "integer", "minimum": 1 }
        }
      },
      "SecuritySettingsRequest": {
        "type": "object",
        "properties": {
          "enforce_strong_password": { "type": "boolean" },
          "password_rotation_days": { "type": "integer", "minimum": 1 },
          "max_session_count": { "type": "integer", "minimum": 1 }
        }
      },
      "AuditExportRequest": {
        "type": "object",
        "properties": {
          "actor_id": { "type": "string", "format": "uuid" },
          "action": { "type": "string" },
          "resource_type": { "type": "string" },
          "resource_id": { "type": "string", "format": "uuid" },
          "start_time": { "type": "string", "format": "date-time" },
          "end_time": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}